+++ src/data/projectData.js
@@ -2,16 +2,15 @@
 export const projectData = [
     {
         name: "Indeed",
-        description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos harum!",
-        image: '../components/ProjectCard/Image/indeedHomepage.jpg'
+        description:  "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos harum. adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium!",
+        image: './src/Image/indeedHomepage.jpg'
     },
     {
         name: "Manifold",
-        description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos harum!"
+        description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor accusantium sapienteveritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos harum!"
     },
     {
         name: "ttv capital",
-        description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos harum!"
+        description: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor veritatis accusamus explicabo eveniet impedit minima accusantium sapiente eius dignissimos haru madi pisicing elit. Dolor veritatis accusamus."
     },
-
   ]
